Updates CollapsibleGroup to Wrap it onOpenHandler
CollapsibleGroup passes down a onOpen handler but there are instances
when the implementer has also passed in an onOpen handler and we don't
want to overwrite that user's onOpen.

If a Collapsible already has an onOpen, instead of overwriting it, we
replace it with a new function that calls the old onOpen and the group's
onOpen handler.